requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
openai==1.2.4
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.0
pyarrow==14.0.1
pydantic==2.5.0
pydantic_core==2.14.1
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
streamlit==1.28.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
flask
aiodns==3.0.0
aleph_alpha_client==3.1.0
apify_client==1.1.1
arize==7.0.1
asynctest==0.13.0
atlassian==0.0.0
banana_dev==4.0.2
beautifulsoup4==4.12.2
bilibili_api==9.1.0
boto3==1.26.131
botocore==1.29.131
Brotli==1.0.9
brotlicffi==1.0.9.2
cchardet==2.1.7
cerebrium==1.1.7
chromadb==0.3.22
clearml==1.10.4
clickhouse_connect==0.5.23
comet_ml==3.33.1
ConfigParser==5.3.0
contextlib2==21.6.0
cryptography==40.0.2
cx_Oracle==8.3.0
Cython==0.29.33
datasets==2.12.0
disco==1.40.4
dl==0.1.0
docarray==0.31.1
docutils==0.20
docx2txt==0.8
duckdb==0.7.1
elasticsearch==8.7.0
email_validator==2.0.0.post2
extract_msg==0.41.0
fake_useragent==1.1.3
firebase_admin==6.1.0
fitz==0.0.1.dev2
genapi==0.0.8
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
googlemaps==4.10.0
gptcache==0.1.22
greenlet==2.0.2
guardrails==2.0.0
gunicorn==20.1.0
HTMLParser==0.0.2
hypothesis==6.75.2
idna_ssl==1.1.0
InstructorEmbedding==1.0.0
ipython==8.13.2
ipywidgets==7.7.0
Jinja2==3.1.2
keyring==23.13.1
lark==1.1.5
lockfile==0.12.2
lxml==4.9.2
matplotlib==3.6.2
metal_sdk==1.0.1
mtrand==0.1
mwparserfromhell==0.6.4
mwxml==0.3.3
networkx==2.8.8
nltk==3.8.1
nomic==1.1.6
nose==1.3.7
ordereddict==1.1
pdf2image==1.16.3
pdfminer==20191125
pexpect==4.8.0
pickle5==0.0.11

pinecone_text==0.4.2
promptlayer==0.1.80
protobuf==4.23.0
pygpt4all==1.1.0
PyInstaller==5.10.1
pymilvus==2.2.8
pyowm==3.3.0
pypandoc==1.11
pypdf==3.8.1
PyPDF2==3.0.1
pyspark==3.4.0
pytz==2022.6
qdrant_client==1.1.7

scipy==1.10.1
selenium==4.9.1
sentence_transformers==2.2.2
setuptools_scm==7.1.0
simplejson==3.19.1
slack_sdk==3.21.3
spacy==3.5.2
Sphinx==7.0.0
sqlcipher3==0.5.0
svglib==1.5.1
tensorflow==2.12.0
tensorflow_text==2.12.1
textstat==0.7.3
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tokio==0.2.0
torch==2.0.1
tornado==6.1
transformers==4.28.1
truststore==0.7.0
tweepy==4.14.0
urllib3_secure_extra==0.1.0
uvloop==0.17.0
vespa==0.6
wandb==0.15.2
wincertstore==0.2.1
wolframalpha==5.0.0
xlrd==2.0.1
xmlrpclib==1.0.1
youtube_transcript_api==0.6.0
zstandard==0.21.0